backend/requirements.txt
chromadb>=0.4.4
openai>=1.7.2
tqdm
llama_index.llms.ollama
flask
flask_socketio
requests